refactor: extract CalendarView from erg-dashboard.jsx
Part 3/3 of issue #78. Moves the inline 'calendar' tab (roster
window, day cells, upcoming events ladder) into a standalone
views/CalendarView.jsx component. erg-dashboard.jsx now has no
inline view logic left except the ErrorBoundary class and routing.
